fix distances and heights in compute enhanced slopes
We want to get one reach upstream and one reach downstream, and connect them togerther to do the Gaussian averaging for node height. If current reach is the 1st reach, we only get the downstream reach. And if the current reach is the last reach, we only grab the upstream reach for smoothing. After smoothing for the connected 2 or 3 reaches, we only get the first and last nodes, and reach length from current reach (not the connected reaches), and then calculated the reach slopes.